Add two new fields to CSV for ingest
Added dc.date.copyright and dcterms.accessRights, which are needed to
track which items are under copyright and to hold a note about usage of
the thesis. These fields are needed now that all records have the same
content in dc.rights and we can no longer use dc.rights to distinguish
the open access items from the restricted access items.